requirements.txt
numpy==1.23.3
pandas==1.5.0
PyYAML==6.0.1
Requests==2.31.0
xmltodict=0.13.0
rdflib==6.3.2
pyparsing==2.1.0
isodate==0.7.0